src/test/java/kr/megaptera/assignment/application/CreatePostServiceTest.java
package kr.megaptera.assignment.application;

import kr.megaptera.assignment.dtos.PostCreateDto;
import kr.megaptera.assignment.dtos.PostDto;
import kr.megaptera.assignment.models.Post;
import kr.megaptera.assignment.repositories.PostRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;

class CreatePostServiceTest {
    private PostRepository postRepository;

    private CreatePostService createPostService;

    @BeforeEach
    void setUp() {
        postRepository = mock(PostRepository.class);

        createPostService = new CreatePostService(postRepository);
    }

    @Test
    @DisplayName("게시물 생성")
    void create() {
        PostCreateDto newPost = new PostCreateDto("title", "author", "test");

        PostDto postDto = createPostService.create(newPost);

        assertThat(postDto.getTitle()).isEqualTo("title");
        assertThat(postDto.getAuthor()).isEqualTo("author");
        assertThat(postDto.getContent()).isEqualTo("test");

        verify(postRepository).save(any(Post.class));
    }
}
